Use latest release for analysis.
Unexpectedly got the following error below. I assume this is because it
is trying to install into the base environment, whereas my tests always
create a separate environment.

CondaUpgradeError: This environment has previously been operated on by a
conda version that's newer
than the conda currently being used. A newer version of conda is
required.
  target environment location: /opt/conda
  current conda version: 4.5.12
  minimum conda version: 4.6